Create README.md
Make a code review and prepare report which describe pros and cons of of delivered code examples. Remember to include your thought and considerations, problems you will have discovered and possible changes or solutions to make that code better.  

The Company INC asked a few independent programmers for to create an simple bank account application which should deliver following functionality: 

Check Account balance 

Deposit income 

Transfer funds to other account 

Close account 

There is also a requirement about how app should look like. It should be a console application with initial menu which has items listed below: 

Create account 

Login in to account 

Exit 

And the second one, after user of that app log into, like: 

Balance 

 Add income 

 Do transfer 

 Close account 

 Log out 

 Exit 

The main security way to authorize users of that app is a credentials (card number + PIN), which should be also generated by app with “Create account” function. This process should implement a Lehn algorithm with assumption that the every card number starts with a sequence of numbers like 400000. The rest of 16 numbered in this sequence should be random generated.